Core FF Polyhedron split off
