+++ Core/HardParticle/Prism.cpp
@@ -3,7 +3,7 @@
 //  BornAgain: simulate and fit scattering at grazing incidence
 //
 //! @file      Core/HardParticle/Prism.cpp
-//! @brief     Implements class IFormFactorPrism.
+//! @brief     Implements class Prism.
 //!
 //! @homepage  http://www.bornagainproject.org
 //! @license   GNU General Public License v3 or higher (see COPYING)
