+++ Core/HardParticle/Prism.h
@@ -3,7 +3,7 @@
 //  BornAgain: simulate and fit scattering at grazing incidence
 //
 //! @file      Core/HardParticle/Prism.h
-//! @brief     Defines class IFormFactorPrism.
+//! @brief     Defines class Prism.
 //!
 //! @homepage  http://www.bornagainproject.org
 //! @license   GNU General Public License v3 or higher (see COPYING)
